Create ccfg.c
This is part of the example code based on CCS, easylink and TI-RTO's RX end. This file is used to configure Boot ROM, start-up code, and SW radio behaviour.
It needs to be opened in CCS, and all code adjustments should match the CC1310 board. This file only shows the underlying code required by CC1310, it is not allowed to copy the code directly into CC1310! To improve the program, it is highly recommended to connect to the CC1310 board and proceed in CCS.